-Mudança em nomes de classes -Criaçao das classes Docente e Credito (falta criar Turma) -Cadastro dos administradores
