+++ src/java/hibernate.cfg.xml
@@ -4,7 +4,7 @@
     <session-factory>
         <property name="hibernate.dialect">org.hibernate.dialect.MySQLDialect</property>
         <property name="hibernate.connection.driver_class">com.mysql.jdbc.Driver</property>
-        <property name="hibernate.connection.url">jdbc:mysql://localhost:3306/Alocacao?zeroDateTimeBehavior=convertToNull</property>
+        <property name="hibernate.connection.url">jdbc:mysql://localhost:3306/Alocacao2?zeroDateTimeBehavior=convertToNull</property>
         <property name="hibernate.connection.username">root</property>
         <property name="hibernate.connection.password">root</property>
         <property name="hibernate.hbm2ddl.auto">update</property>
@@ -12,10 +12,12 @@
         <property name="hibernate.current_session_context_class">thread</property>
         <property name="hibernate.current_session_context_class">org.hibernate.context.internal.ThreadLocalSessionContext</property>
     
-        <mapping class="model.Afinidades"/>
+        <mapping class="model.Afinidade"/>
         <mapping class="model.Pessoa"/>
+        <mapping class="model.Credito"/>
+        <mapping class="model.Docente"/>
         <mapping class="model.Disciplina"/>
-        <mapping class="model.TurmasPlanejamento"/>
+        <mapping class="model.OfertaDisciplina"/>
         <mapping class="model.Disponibilidade"/>
 
     </session-factory>
