--- src/java/model/Horario.java
@@ -1,14 +1,0 @@
-/*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package model;
-
-/**
- *
- * @author vinivcp
- */
-class Horario {
-    
-}
