+++ src/java/model/Pessoa.java
@@ -4,18 +4,22 @@ import java.io.Serializable;
 import java.util.Set;
 import javax.persistence.CascadeType;
 import javax.persistence.Column;
+import javax.persistence.DiscriminatorColumn;
+import javax.persistence.DiscriminatorType;
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
+import javax.persistence.Inheritance;
+import javax.persistence.InheritanceType;
 import javax.persistence.OneToMany;
 
-
 @Entity
-public class Pessoa implements Serializable {
+@Inheritance(strategy = InheritanceType.SINGLE_TABLE)
+@DiscriminatorColumn(name = "PESSOA_DETAILS_TYPE",discriminatorType = DiscriminatorType.STRING)
+public class Pessoa implements Serializable, Comparable<Pessoa> {
 
-    
     private static final long serialVersionUID = 1L;
 
     public Pessoa() {
@@ -24,12 +28,8 @@ public class Pessoa implements Serializable {
 
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
-    @Column(name="pessoa_id")
+    @Column(name = "pessoa_id")
     private Long ID;
-
-    Pessoa(Long pessoaId) {
-        throw new UnsupportedOperationException("Not supported yet."); //To change body of generated methods, choose Tools | Templates.
-    }
 
     public Long getID() {
         return ID;
@@ -48,12 +48,15 @@ public class Pessoa implements Serializable {
     public void setNome(String nome) {
         this.nome = nome;
     }
-    
+
     private String siape;
-    
+
     private String email;
-    
+
     private String centro;
+
+    private boolean adm;
+    
 
     public String getSiape() {
         return siape;
@@ -78,26 +81,32 @@ public class Pessoa implements Serializable {
     public void setCentro(String centro) {
         this.centro = centro;
     }
+
+    public boolean isAdm() {
+        return adm;
+    }
+
+    public void setAdm(boolean adm) {
+        this.adm = adm;
+    }
     
-    
-    @OneToMany(fetch = FetchType.EAGER, mappedBy = "pessoa", cascade = CascadeType.ALL)
-//            ,cascade = CascadeType.ALL)
-//    @Cascade(value = { org.hibernate.annotations.CascadeType.ALL })  
-//    @LazyCollection(value = LazyCollectionOption.EXTRA)  
-//    @Fetch(value = FetchMode.SUBSELECT) 
-    private Set<Afinidades> afinidades;
     
 
-    public Set<Afinidades> getAfinidades() {
+    @OneToMany(fetch = FetchType.EAGER, mappedBy = "pessoa", cascade = CascadeType.ALL)
+    private Set<Afinidade> afinidades;
+
+    public Set<Afinidade> getAfinidades() {
         return afinidades;
     }
 
-    public void setAfinidades(Set<Afinidades> afinidades) {
+    public void setAfinidades(Set<Afinidade> afinidades) {
         this.afinidades = afinidades;
     }
-    
+
     @OneToMany(fetch = FetchType.EAGER, mappedBy = "pessoa", cascade = CascadeType.ALL)
     private Set<Disponibilidade> disponibilidades;
+    
+    
 
     public Set<Disponibilidade> getDisponibilidades() {
         return disponibilidades;
@@ -106,6 +115,18 @@ public class Pessoa implements Serializable {
     public void setDisponibilidades(Set<Disponibilidade> disponibilidades) {
         this.disponibilidades = disponibilidades;
     }
+
+    @OneToMany(fetch = FetchType.EAGER, cascade = CascadeType.ALL)
+    private Set<Credito> creditos;
+
+    public Set<Credito> getCreditos() {
+        return creditos;
+    }
+
+    public void setCreditos(Set<Credito> creditos) {
+        this.creditos = creditos;
+    }
+    
     
 
     @Override
@@ -135,6 +156,12 @@ public class Pessoa implements Serializable {
     @Override
     public String toString() {
         return this.nome;
+    }
+    
+    @Override
+    public int compareTo(Pessoa o) {
+    
+        return this.getNome().compareTo(o.getNome());
     }
 
 }
