+++ web/Disponibilidade/FaseIQuad2.xhtml
@@ -132,8 +132,8 @@
                                 <p:column headerText="Câmpus" style="width: 6%">
                                     <h:outputText value="#{turma.campus}" style="width: 6%"/>
                                 </p:column>
-                                <p:column headerText="Quad" style="width: 6%">
-                                    <h:outputText value="#{turma.quadrimestre}" style="width: 6%"/>
+                                <p:column headerText="Qdt. turmas" style="width: 8%">
+                                    <h:outputText value="#{turma.numTurmas}" style="width: 6%"/>
                                 </p:column>
                             </p:dataTable>
